chat-app/requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.1.4
decorator==4.1.2
eventlet==0.17.4
filelock==3.0.12
Flask==0.12.2
flask-htpasswd==0.3.1
Flask-Login==0.4.1
Flask-SocketIO==2.9.3
Flask-WTF==0.14.2
gevent==1.4.0
greenlet==0.4.15
idna==2.6
importlib-metadata==0.23
importlib-resources==1.0.2
infinity==1.4
intervals==0.8.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
nltk==3.2.5
packaging==20.1
passlib==1.7.2
pathlib2==2.3.5
pluggy==0.13.1
py==1.8.1
pycparser==2.18
pyparsing==2.4.6
python-engineio==2.0.1
python-socketio==1.8.4
requests==2.18.4
scandir==1.10.0
schedule==0.1.10
six==1.14.0
toml==0.10.0
tox==3.14.0
typing==3.7.4.1
urllib3==1.22
validators==0.12.0
virtualenv==20.0.1
w3lib==1.18.0
websocket==0.2.1
websocket-client==0.57.0
websocket-server==0.4
Werkzeug==0.13
WTForms==2.1
WTForms-Components==0.10.3
wtforms-html5==0.2.2
zipp==1.1.0
